Servo adjustment remote control
This software turns a Wemos D1 mini into a remote control to adjust the
servo of a turnout.
The address of the servo is selected by means of dip-switches. The
position of a potmeter is mapped to angles. This position is published.
The straight position is acknowledged with a push button.
The thrown position is acknowledged with a push button.
Both push button actions are published.
The process starts again for the next servo.
In total 64 servos can be adjusted with this version of the software.